feat(blocks): support custom placeholder for paragraph (#6787)
Close BS-262

Need to config it in affine side. cc @pengx17
