+++ packages/blocks/src/paragraph-block/paragraph-service.ts
@@ -26,4 +26,21 @@ export class ParagraphService<
     this.inlineManager.registerSpecs(inlineSpecs);
     this.inlineManager.registerMarkdownMatches(affineInlineMarkdownMatches);
   }
+
+  placeholderGenerator: (model: ParagraphBlockModel) => string = model => {
+    if (model.type === 'text') {
+      return "Type '/' for commands";
+    }
+
+    const placeholders = {
+      h1: 'Heading 1',
+      h2: 'Heading 2',
+      h3: 'Heading 3',
+      h4: 'Heading 4',
+      h5: 'Heading 5',
+      h6: 'Heading 6',
+      quote: '',
+    };
+    return placeholders[model.type];
+  };
 }
