+++ packages/presets/src/ai/ai.ts
@@ -11,7 +11,9 @@ import {
   EdgelessEditorBlockSpecs,
   EdgelessElementToolbarWidget,
   PageEditorBlockSpecs,
+  ParagraphService,
 } from '@blocksuite/blocks';
+import { assertInstanceOf } from '@blocksuite/global/utils';
 import { literal, unsafeStatic } from 'lit/static-html.js';
 
 import { buildAIPanelConfig } from './ai-panel.js';
@@ -58,6 +60,36 @@ export function patchDocSpecs(specs: BlockSpec[]) {
       };
       return newPageSpec;
     }
+
+    if (spec.schema.model.flavour === 'affine:paragraph') {
+      const newParagraphSpec: BlockSpec = {
+        ...spec,
+        setup(slots, disposableGroup) {
+          spec.setup?.(slots, disposableGroup);
+          slots.mounted.on(({ service }) => {
+            assertInstanceOf(service, ParagraphService);
+            service.placeholderGenerator = model => {
+              if (model.type === 'text') {
+                return "Type '/' for commands, 'space' for AI";
+              }
+
+              const placeholders = {
+                h1: 'Heading 1',
+                h2: 'Heading 2',
+                h3: 'Heading 3',
+                h4: 'Heading 4',
+                h5: 'Heading 5',
+                h6: 'Heading 6',
+                quote: '',
+              };
+              return placeholders[model.type];
+            };
+          });
+        },
+      };
+      return newParagraphSpec;
+    }
+
     return spec;
   });
 }
